requirements.txt
# TerraSeg Inference Pipeline Requirements
# Install with: pip install -r requirements.txt

torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.3
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.0.0
tqdm>=4.65.0
gdown>=4.7.0
